more optimisations

+++ src/Controller/Inputs.ino
@@ -75,7 +75,7 @@ void initInput() {
 }
 
 void parseInputs() {
-  int old_state;
+  byte old_state;
 
   // faders
